Test random with test (from minitest!)
Due to the stateful nature of a seeded random number generator, testing
the random standard library the way most of the other libspw modules are
tested (executable stack effects in the comments) would've gotten messy
very quickly. Everything is in place to feed arbitrary instructions into
the hot-loadable Whitespace interpreter used for testing, but going that
route would've required coming up with some (probably unpleasant) syntax
for including instructions amongst the stack effects.

Instead, Spitewaste's test library gets to shine! We execute a program
that invokes the functions from the random library and asserts that the
outputs match expectations. `assert_eq` dies if some equality doesn't
hold, which in turn ensures that we don't exit with a clean stack, and
a non-empty exit stack will cause the minitest assertion to fail.